feat(cmd): Rewrite elastic status to make it reusable
